backend/pyproject.toml
[project]
name = "stash-backend"
version = "1.0.0"
description = "Voice-Native AI Supply Chain Platform - Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.18",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "apscheduler>=3.10.0",
    "httpx>=0.28.0",
    "google-generativeai>=0.8.0",
    "google-cloud-speech>=2.28.0",
    "google-cloud-texttospeech>=2.21.0",
    "google-cloud-bigquery>=3.27.0",
    "google-cloud-storage>=2.18.0",
    "google-cloud-pubsub>=2.27.0",
    "twilio>=9.4.0",
    "gTTS>=2.5.0",
    "weasyprint>=63.0",
    "jinja2>=3.1.0",
    "geoalchemy2>=0.15.0",
    "python-dateutil>=2.9.0",
    "numpy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]
